Refactors `cloud/application.py` file and turns it into a directory
