+++ docs/nextmv/cloud/reference/application.md
@@ -2,4 +2,50 @@
 
 This section documents the application components of the Nextmv Cloud API.
 
+The application module is organized with the main `Application` class and
+several mixin classes that provide specialized functionality. Every method
+belonging to a mixin class can be accessed through an instance of the `Application`.
+
+## Main Class
+
 ::: nextmv.nextmv.cloud.application
+
+## Mixin Classes
+
+These mixins extend the `Application` class with specific capabilities.
+
+### Acceptance Testing
+
+::: nextmv.nextmv.cloud.application._acceptance
+
+### Batch Scenarios
+
+::: nextmv.nextmv.cloud.application._batch_scenario
+
+### Ensemble Management
+
+::: nextmv.nextmv.cloud.application._ensemble
+
+### Input Sets
+
+::: nextmv.nextmv.cloud.application._input_set
+
+### Instance Management
+
+::: nextmv.nextmv.cloud.application._instance
+
+### Managed Inputs
+
+::: nextmv.nextmv.cloud.application._managed_input
+
+### Run Management
+
+::: nextmv.nextmv.cloud.application._run
+
+### Secrets Management
+
+::: nextmv.nextmv.cloud.application._secrets
+
+### Version Management
+
+::: nextmv.nextmv.cloud.application._version
