+++ docs/nextmv/cloud/runs.md
@@ -380,13 +380,13 @@ Run ID: latest-rq0pw6sy, Status: succeeded
 [cancel-run-section]: #cancel-a-run
 [running-non-json-section]: #running-with-non-json-payloads
 [webhooks]: https://docs.nextmv.io/docs/using-nextmv/setup/webhooks
-[app-new-run]: ./reference/application.md#nextmv.nextmv.cloud.application.Application.new_run
-[app-run-metadata]: ./reference/application.md#nextmv.nextmv.cloud.application.Application.run_metadata
-[app-list-runs]: ./reference/application.md#nextmv.nextmv.cloud.application.Application.list_runs
-[app-run-result]: ./reference/application.md#nextmv.nextmv.cloud.application.Application.run_result
-[app-run-result-with-polling]: ./reference/application.md#nextmv.nextmv.cloud.application.Application.run_result_with_polling
-[app-new-run-with-result]: ./reference/application.md#nextmv.nextmv.cloud.application.Application.new_run_with_result
-[app-cancel-run]: ./reference/application.md#nextmv.nextmv.cloud.application.Application.cancel_run
+[app-new-run]: ./reference/application.md#nextmv.nextmv.cloud.application._run.ApplicationRunMixin.new_run
+[app-run-metadata]: ./reference/application.md#nextmv.nextmv.cloud.application._run.ApplicationRunMixin.run_metadata
+[app-list-runs]: ./reference/application.md#nextmv.nextmv.cloud.application._run.ApplicationRunMixin.list_runs
+[app-run-result]: ./reference/application.md#nextmv.nextmv.cloud.application._run.ApplicationRunMixin.run_result
+[app-run-result-with-polling]: ./reference/application.md#nextmv.nextmv.cloud.application._run.ApplicationRunMixin.run_result_with_polling
+[app-new-run-with-result]: ./reference/application.md#nextmv.nextmv.cloud.application._run.ApplicationRunMixin.new_run_with_result
+[app-cancel-run]: ./reference/application.md#nextmv.nextmv.cloud.application._run.ApplicationRunMixin.cancel_run
 [cancel-a-run]: #cancel-a-run
 [input]: ../modeling/reference/input.md#nextmv.nextmv.input.Input
 [inputformat]: ../modeling/reference/input.md#nextmv.nextmv.input.InputFormat
